pyproject.toml
[tool.poetry]
name = "cortex"
version = "0.1.0"
description = "Cortex Mentor Application"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.111.0"
uvicorn = {version = "0.29.0", extras = ["standard"]}
pydantic = "2.7.1"
chromadb = "0.5.0"
ollama = "0.2.0"
watchdog = "4.0.0"
GitPython = "3.1.43"
arq = "^0.26.0"
redis = "^5.0.4"
upstash-vector = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
